fix routing and unnecessary duping
