--- Open/pages/qualifier/index.vue
@@ -1,23 +1,0 @@
-<template>
-    <div />
-</template>
-
-<script lang="ts">
-import { Vue, Component } from "vue-property-decorator";
-
-@Component({
-    components: {
-    },
-    head () {
-        return {
-            title: "Corsace Open",
-        };
-    },
-})
-export default class Qualifier extends Vue {
-
-}
-</script>
-
-<style lang="scss">
-</style>
